+++ Net Pools/src/ValueObjectWrapper.cs
@@ -6,26 +6,28 @@ namespace Enderlook.Pools
 {
     internal struct ValueObjectWrapper<T> where T : struct
     {
+        // UNLOCK_EMPTY = 0;
+        // LOCK = 1;
+        // UNLOCK_FULL = 2;
         private T value;
-        private bool hasValue;
         private int @lock;
 
-        public bool HasValue
+        public bool NotSynchronizedHasValue
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => hasValue;
+            get => @lock == 2;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool TryPopValue([NotNullWhen(true)] out T value)
         {
-            while (Interlocked.Exchange(ref @lock, 1) == 1) ;
+            int oldLock;
+            while ((oldLock = Interlocked.Exchange(ref @lock, 1)) == 1) ;
 
-            if (hasValue)
+            if (oldLock == 2)
             {
                 value = this.value;
                 this.value = default;
-                hasValue = false;
                 @lock = 0;
                 return true;
             }
@@ -42,17 +44,17 @@ namespace Enderlook.Pools
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool TrySetValue(ref T value)
         {
-            while (Interlocked.Exchange(ref @lock, 1) == 1) ;
+            int oldLock;
+            while ((oldLock = Interlocked.Exchange(ref @lock, 1)) == 1) ;
 
-            if (hasValue)
+            if (oldLock == 2)
             {
-                @lock = 0;
+                @lock = 2;
                 return false;
             }
 
             this.value = value;
-            hasValue = true;
-            @lock = 0;
+            @lock = 2;
             return true;
         }
 
@@ -60,7 +62,6 @@ namespace Enderlook.Pools
         public void Clear()
         {
             while (Interlocked.Exchange(ref @lock, 1) == 1) ;
-            hasValue = false;
             value = default;
             @lock = 0;
         }