+++ Net Pools/src/CompilationHelper.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Reflection.Emit;
@@ -7,7 +6,7 @@ using System.Runtime.CompilerServices;
 
 namespace Enderlook.Pools
 {
-    internal static class Utilities
+    internal static class CompilationHelper
     {
         public const int DisabledDynamicCompilation = 0;
         public const int SystemLinqExpressions = 1;
@@ -16,7 +15,7 @@ namespace Enderlook.Pools
 
         public static readonly ParameterExpression[] EmptyParameters = new ParameterExpression[0];
 
-        static Utilities()
+        static CompilationHelper()
         {
             const BindingFlags flags = BindingFlags.Public | BindingFlags.Static;
 
@@ -62,7 +61,7 @@ namespace Enderlook.Pools
 #if NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
             try
             {
-                DynamicMethod dynamicMethod = new ("Instantiate", typeof(object), Type.EmptyTypes);
+                DynamicMethod dynamicMethod = new("Instantiate", typeof(object), Type.EmptyTypes);
                 ILGenerator generator = dynamicMethod.GetILGenerator();
                 generator.Emit(OpCodes.Newobj, typeof(object).GetConstructor(Type.EmptyTypes)!);
                 generator.Emit(OpCodes.Ret);
@@ -80,45 +79,5 @@ namespace Enderlook.Pools
             isDisabled:
             DynamicCompilationMode = DisabledDynamicCompilation;
         }
-
-        public enum MemoryPressure
-        {
-            Low,
-            Medium,
-            High
-        }
-
-        public static MemoryPressure GetMemoryPressure()
-        {
-#if NET5_0_OR_GREATER
-            const double HighPressureThreshold = .90; // Percent of GC memory pressure threshold we consider "high".
-            const double MediumPressureThreshold = .70; // Percent of GC memory pressure threshold we consider "medium".
-
-            GCMemoryInfo memoryInfo = GC.GetGCMemoryInfo();
-
-            if (memoryInfo.MemoryLoadBytes >= memoryInfo.HighMemoryLoadThresholdBytes * HighPressureThreshold)
-                return MemoryPressure.High;
-
-            if (memoryInfo.MemoryLoadBytes >= memoryInfo.HighMemoryLoadThresholdBytes * MediumPressureThreshold)
-                return MemoryPressure.Medium;
-
-            return MemoryPressure.Low;
-#else
-            return MemoryPressure.High;
-#endif
-        }
-
-        [DoesNotReturn]
-        public static void ThrowArgumentNullException_Obj()
-            => throw new ArgumentNullException("obj");
-
-        public static void ThrowArgumentOutOfRangeException_InitialColdCapacityCanNotBeNegative()
-            => throw new ArgumentOutOfRangeException("initialColdCapacity", "Can't be negative.");
-
-        public static void ThrowArgumentOutOfRangeException_InitialCapacityCanNotBeNegative()
-            => throw new ArgumentOutOfRangeException("initialCapacity", "Can't be negative.");
-
-        public static void ThrowArgumentOutOfRangeException_HotCapacityCanNotBeLowerThanOne()
-            => throw new ArgumentOutOfRangeException("hotCapacity", "Can't be lower than 1.");
     }
 }