+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fix documentation typos.
 - Fix error when using value types without parameterless constructor.
 - Reduce memory consumption of `DynamicValueObjectPool<T>`.
+- Minimal performance improvements.
 
 ## v0.2.2
 - Add support for trimming.